Greatly improve trip matching performance; make trip mappings persistent in EHCache.
We now look for trips which are active 10 minutes after the trip start time we get from the WMATA API, to account for trips in the GTFS that haven't started yet.  We also exclude stoptime mappings with very large scores, to keep one bad stoptime from preventing the whole trip from being mapped.